[#126584711] Use WelcomePresenter to pass statistics to welcome_view, replacing the helpers that were previously used
