+++ Gemfile
@@ -55,6 +55,7 @@ group :development, :test do
   gem "spring", "1.4.3"
   gem "rubocop", require: false
   gem "letter_opener"
+  gem 'selenium-webdriver', '~> 2.53', '>= 2.53.4'
 end
 
 group :development do
