+++ app/controllers/welcome_controller.rb
@@ -16,6 +16,7 @@ class WelcomeController < ApplicationController
   end
 
   def about
+    @presenter = Welcome::WelcomePresenter.new
   end
 
   def renders(events)
