--- app/helpers/welcome_helper.rb
@@ -1,15 +1,0 @@
-module WelcomeHelper
-
-  def attendee_count
-    Attendee.count
-  end
-
-  def ticket_count
-    UserTicket.count
-  end
-
-  def event_count
-    Event.count
-  end
-  
-end
