+++ spec/features/attending_events_spec.rb
@@ -37,7 +37,7 @@ RSpec.feature "ViewEvents", type: :feature, js: true do
     expect(page).to have_content "ADD TO GOOGLE CALENDAR"
   end
 
-  scenario "User clicks to attend an event" do
+  scenario "User clicks to attend an event", :js => true do
     sign_up
     visit events_path
     click_link "Blessings wedding"
